+++ lib/canvas/src/shapes/shape.dart
@@ -1,0 +1,43 @@
+import 'dart:ui';
+import 'package:countries_world_map/canvas/touchable_canvas.dart';
+import 'package:flutter/gestures.dart';
+import 'package:flutter/material.dart';
+
+class ShapeConstant {
+  static const floatPrecision = 0.001;
+  static const double infinity = 9999999;
+}
+
+abstract class Shape {
+  final Paint paint;
+  final Map<GestureType, Function> gestureCallbackMap;
+  final HitTestBehavior hitTestBehavior;
+
+  Set<GestureType> get registeredGestures => gestureCallbackMap.keys.toSet();
+
+  Shape({
+    Paint? paint,
+    Map<GestureType, Function>? gestureCallbackMap,
+    HitTestBehavior? hitTestBehavior,
+  })  : paint = paint ??
+            (Paint()
+              ..strokeWidth = ShapeConstant.floatPrecision
+              ..style = PaintingStyle.fill),
+        gestureCallbackMap = gestureCallbackMap ?? {},
+        hitTestBehavior = hitTestBehavior ?? HitTestBehavior.opaque {
+    if (this.paint.strokeWidth == 0) {
+      this.paint.strokeWidth = ShapeConstant.floatPrecision;
+    }
+  }
+
+  bool isInside(Offset p);
+
+  Function getCallbackFromGesture(Gesture gesture) {
+    if (gestureCallbackMap.containsKey(gesture.gestureType)) {
+      return () =>
+          gestureCallbackMap[gesture.gestureType]?.call(gesture.gestureDetail);
+    } else {
+      return () {};
+    }
+  }
+}