+++ lib/countries/europe/netherlands/netherlands.dart
@@ -1,0 +1,5 @@
+library netherlands;
+
+export 'src/netherlands_colors.dart';
+export 'src/netherlands_map.dart';
+export 'src/netherlands_painter.dart';