+++ lib/countries/europe/netherlands/src/netherlands_colors.dart
@@ -1,0 +1,67 @@
+import 'dart:ui';
+
+/// This class takes in a color for each province in the Netherlands.
+///
+/// - [groningen]
+/// - [friesland]
+/// - [drenthe]
+/// - [overijssel]
+/// - [flevoland]
+/// - [gelderland]
+/// - [utrecht]
+/// - [noordholland]
+/// - [zuidholland]
+/// - [zeeland]
+/// - [noordbrabant]
+/// - [limburg]
+class NetherlandsColors {
+  /// Groningen
+  final Color? groningen;
+
+  /// Friesland
+  final Color? friesland;
+
+  /// Drenthe
+  final Color? drenthe;
+
+  /// Overijssel
+  final Color? overijssel;
+
+  /// Flevoland
+  final Color? flevoland;
+
+  /// Gelderland
+  final Color? gelderland;
+
+  /// Utrecht
+  final Color? utrecht;
+
+  /// Noord Holland
+  final Color? noordholland;
+
+  /// Zuid Holland
+  final Color? zuidholland;
+
+  /// Zeeland
+  final Color? zeeland;
+
+  /// Noord Brabant
+  final Color? noordbrabant;
+
+  /// Limburg
+  final Color? limburg;
+
+  const NetherlandsColors(
+      {this.groningen,
+      this.friesland,
+      this.drenthe,
+      this.overijssel,
+      this.flevoland,
+      this.gelderland,
+      this.utrecht,
+      this.noordholland,
+      this.zuidholland,
+      this.zeeland,
+      this.noordbrabant,
+      this.limburg});
+}