+++ lib/world/simple_world/simple_world.dart
@@ -1,0 +1,5 @@
+library simple_world_map;
+
+export 'src/simple_world_colors.dart';
+export 'src/simple_world_map.dart';
+export 'src/simple_world_painter.dart';